프레젠테이션 유틸을 공용 모듈로 통합하고 import 가드 추가
